*improved parser for 1->n relationshipts over non book fields * made update progress bar  force show text by default * modified ordered set so for non existing elements it returns -1 (didn't feel ok with None) + added separate class for titles + added dictionary options for the create functions (useful for debugging, but I guess it will remain like this for production)

+++ gutenbergpy/gutenbergcache.py
@@ -6,48 +6,39 @@ from gutenbergpy.gutenbergcachesettings import GutenbergCacheSettings
 from gutenbergpy.parse.rdfparser        import RdfParser
 from gutenbergpy.caches.sqlitecache      import SQLiteCache
 
-'''  'publisher':ParseTemp(tablename='publishers',xpath='//dcterms:publisher/text()'),
-    'description':ParseTemp(tablename='description',xpath='//dcterms:description/text()'),'''
-
 
 class GutenbergCache:
 
     @staticmethod
-    def create(refresh = False):
+    def create(**kwargs):
 
-        if path.isfile(GutenbergCacheSettings.CACHE_FILENAME) and refresh is False:
+        if path.isfile(GutenbergCacheSettings.CACHE_FILENAME) and kwargs['refresh'] == True :
             print 'Cache already exists'
             return
 
-        if refresh is True:
+        if kwargs['refresh'] == True:
             print 'Deleting old files'
             Utils.delete_tmp_files(True)
 
-        Utils.download_file()
-        Utils.unpack_tarbz2()
+        if kwargs['download'] == True:
+            Utils.download_file()
 
-        t0 = time.time()
-        parser = RdfParser()
-        result = parser.do()
-        print 'RDF PARSING took ' + str(time.time() - t0)
+        if kwargs['unpack'] == True:
+            Utils.unpack_tarbz2()
 
-        t0 = time.time()
-        cache = SQLiteCache()
-        cache.create_cache(result)
-        print 'sql took %f' % (time.time() - t0)
+        if kwargs['parse'] == True:
+            t0 = time.time()
+            parser = RdfParser()
+            result = parser.do()
+            print 'RDF PARSING took ' + str(time.time() - t0)
 
+        if kwargs['cache'] == True:
+            t0 = time.time()
+            cache = SQLiteCache()
+            cache.create_cache(result)
+            print 'sql took %f' % (time.time() - t0)
 
-        print 'Deleting temporary files'
-        GutenbergCache.delete_tmp_files(True)
+        if kwargs['deleteTemp'] == True:
+            print 'Deleting temporary files'
+            GutenbergCache.delete_tmp_files(True)
         print 'Done'
-
-
-
-''' query english books
-select titles.name as booktitle, languages.name as language from titles, books,book_titles,book_languages, languages where
-titles.id = book_titles.title_id and
-books.id = book_titles.book_id and
-book_languages.bookid = books.id and
-book_languages.languageid = languages.id and
-languages.name = "en"
-'''
